requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
idna==3.10
newsapi-python==0.2.7
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.3
